+++ client/src/extension.ts
@@ -38,7 +38,10 @@ export function activate(context: ExtensionContext) {
 	// Options to control the language client
 	const clientOptions: LanguageClientOptions = {
 		// Register the server for plain text documents
-		documentSelector: [{ scheme: 'file', language: 'inform' }],
+		documentSelector: [
+			{ scheme: 'file', language: 'inform' },
+			{ scheme: 'file', language: 'generaldat' }
+		],
 		synchronize: {
 			// Notify the server about file changes to '.clientrc files contained in the workspace
 			fileEvents: workspace.createFileSystemWatcher('**/.clientrc')